* Fixed the playlist to continue playing
The most significant changes include the update of the `MediaElement` in `MainPage.xaml` to include a `MediaEnded` event handler, the addition of the `Microsoft.Maui.Controls.Internals` namespace to `MainPage.xaml.cs`, and the update of the `PlaylistContext` to be passed to the `MainPage` constructor. The `_mainSongList` is now populated with songs from the database and sorted alphabetically. The logic for setting the `_mediaSource` has been updated to handle embedded resources.

The `TestSonglist_ItemSelected` method has been updated to handle embedded resources and provide more detailed debug output. A `mediaElement_MediaEnded` method has been added to handle the end of a media file and automatically start playing the next song in the list. The `Song` objects added to the database in `MauiProgram.cs` have been updated to use embedded resources and have more detailed names.

The README.md file has been updated to reflect the project's progress, including the ability to play more than one song from a static playlist and the discovery of a useful code snippet. The Dev Blog section of the README.md file has been updated with recent developments and discoveries.

List of changes:

1. Update of `MediaElement` in `MainPage.xaml` to include a `MediaEnded` event handler.
2. Addition of `Microsoft.Maui.Controls.Internals` namespace to `MainPage.xaml.cs`.
3. Update of `PlaylistContext` to be passed to the `MainPage` constructor.
4. Population and sorting of `_mainSongList` with songs from the database.
5. Update of `_mediaSource` setting logic to handle embedded resources.
6. Update of `TestSonglist_ItemSelected` method to handle embedded resources and provide more detailed debug output.
7. Addition of `mediaElement_MediaEnded` method to handle the end of a media file and automatically start playing the next song.
8. Update of `Song` objects in `MauiProgram.cs` to use embedded resources and have more detailed names.
9. Update of README.md file to reflect project progress and recent discoveries.
10. Update of Dev Blog section in README.md file with recent developments and discoveries.