* 240221 - Completed the first iteration of the Log Viewer Window.
This is probably not the way to implement this, but I wanted to finish this path to learn everything I could from it.

The most significant changes in the code include the modification of the `Dispatcher` in `MainPage.xaml.cs`. Additionally, the `ListView` name was changed in `MyPage.xaml`, a new `ObservableCollection` was created in `MyPage.xaml.cs`, and notes about the changes were added to `README.md`.

Here is a list of the changes:

The code for creating a second window was removed from `MainPage.xaml.cs` and replaced with a call to a new method `SecondWindow()`. The `Dispatcher` was also changed from `Application.Current.MainPage.Dispatcher.Dispatch` to `this.Dispatcher.Dispatch`. This change makes it easier to read, and simplifies the process of creating a second window and makes the `Dispatcher` more specific to the current instance (MainPage.xaml.cs).

The name of the `ListView` was changed from `Debuglist` to `DataWindow` in `MyPage.xaml`. This change makes the name of the `ListView` more descriptive and relevant to its function (MyPage.xaml).

A new `ObservableCollection` named `pageOfData` was created in `MyPage.xaml.cs` to hold the data. The code for reading from a file and updating the `ListView` was also modified to use `this.Dispatcher.DispatchAsync` and `ObservableCollection`. A new method `SecondWindow()` was added to create and open a second window. These changes improve the data handling and window creation processes (MyPage.xaml.cs).

Notes about the changes made in the code were added to `README.md`. These notes provide documentation for the changes and explain the use of `ObservableCollection`, `Dispatcher.DispatchAsync`, and `FileShare.ReadWrite`, as well as the creation of a second window (README.md).