*** Meet Version 0.1
The most significant changes include the addition of a new constant `InternalVersion` to the `Const` class in the `AhConfig` namespace, the update of the `queuedMsg` string to include the `InternalVersion` constant, and the extensive update of the To-Do list in the `README.md` file. A new Dev Blog entry was also added to the `README.md` file.

1. A new constant `InternalVersion` was added to the `Const` class in the `AhConfig` namespace in the `Constants.cs` file. This constant holds the internal version of the app. (See `Constants.cs`)

2. The `queuedMsg` string in the `MauiMediaPlayer.ProgramLogic` namespace in the `StaticProgramLogic.cs` file was updated to include the `InternalVersion` constant. (See `StaticProgramLogic.cs`)

3. The To-Do list in the `README.md` file was updated. Tasks related to Delete Operation Optimization, Search For Music Separate DbContext with No-Tracking, and Fix GitHub insisting on messing up lowercase file names were removed. New tasks related to API work, refactoring, code cleanup, and the Repository Pattern and Interfaces were added. (See `README.md`)