Working on Sort Order, in preparation for Playlist Order.  The database is not saving list by order, but instead is in song.id order.
The most significant changes include the modification of the `DispatchSonglist` function in the `MauiMediaPlayer` namespace to include an optional `_searchBy` parameter and sorting functionality based on this parameter. The `ChangePlaylist` function in the same namespace was also updated to accept an integer ID instead of a Playlist object. The `MainPage.xaml` file was updated to disable the `StandardSearchBar` and to change the margins of the `TestSonglist` and `SongGrid`. Lastly, the `MainPage.xaml.cs` file was updated to remove the sorting of `_songList` by `AlphaTitle` in the `LoadDatabase` and `ReloadDatabase` functions.

List of changes:

1. The minimum log level in the `AhConfig` namespace was changed from 2 to 1.
2. The search parameters in the `DataLibrary` namespace were updated to include "Playlist".
3. The `DispatchSonglist` function in the `MauiMediaPlayer` namespace was updated to include an optional `_searchBy` parameter and sorting functionality based on this parameter.
4. The `ChangePlaylist` function in the `MauiMediaPlayer` namespace was updated to accept an integer ID instead of a Playlist object.
5. The `AddSongsGui_Clicked` function in the `MauiMediaPlayer` namespace was updated to include a comment change.
6. The `MainPage.xaml` file was updated to disable the `StandardSearchBar` and to change the margins of the `TestSonglist` and `SongGrid`.
7. The `MainPage.xaml.cs` file was updated to remove the sorting of `_songList` by `AlphaTitle` in the `LoadDatabase` and `ReloadDatabase` functions.
8. The README.md file was updated to include a note about working on Sort Order and to add a new entry in the Dev Blog about layout stability issues.